add Split method
